@@ -1,9 +1,10 @@
-module solapi-agent
+module github.com/bck-newsalt/solapi-agent
 
-go 1.18
+go 1.22.0
 
 require (
 	github.com/go-sql-driver/mysql v1.6.0
+	github.com/jackc/pgx/v5 v5.5.3
 	github.com/solapi/solapi-go v0.0.0-20210702030642-9c9f1d2ee470
 	github.com/takama/daemon v1.0.0
 	gopkg.in/natefinch/lumberjack.v2 v2.0.0
@@ -11,6 +12,10 @@ require (
 
 require (
 	github.com/BurntSushi/toml v0.3.1 // indirect
-	golang.org/x/sys v0.0.0-20220513210249-45d2b4557a2a // indirect
+	github.com/jackc/pgpassfile v1.0.0 // indirect
+	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
+	golang.org/x/crypto v0.17.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
