@@ -1,11 +1,8 @@
 package main
 
 import (
-	"database/sql"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"log"
 	"os"
 	"os/signal"
 	"runtime"
@@ -15,21 +12,11 @@ import (
 	"time"
 	"unsafe"
 
-	_ "github.com/go-sql-driver/mysql"
+	"github.com/bck-newsalt/solapi-agent/cmd/database"
+	logger "github.com/bck-newsalt/solapi-agent/cmd/logger"
 	"github.com/solapi/solapi-go"
 	"github.com/takama/daemon"
-	"gopkg.in/natefinch/lumberjack.v2"
 )
-
-type DBConfig struct {
-	Provider string `json:"provider"`
-	DBName   string `json:"dbname"`
-	Table    string `json:"table"`
-	User     string `json:"user"`
-	Password string `json:"password"`
-	Host     string `json:"host"`
-	Port     int    `json:"port"`
-}
 
 type APIConfig struct {
 	APIKey          string `json:"apiKey"`
@@ -46,16 +33,11 @@ const (
 	description = "Solapi Agent Service"
 )
 
-var dbconf DBConfig
 var apiconf APIConfig
-
-var stdlog, errlog *log.Logger
 
 var client *solapi.Client
 
-var db *sql.DB
-
-var homedir string = "/opt/agent"
+var basePath string = "/opt/agent"
 
 // Service has embedded daemon
 type Service struct {
@@ -92,20 +74,17 @@ func (service *Service) Manage() (string, error) {
 	interrupt := make(chan os.Signal, 1)
 	signal.Notify(interrupt, os.Interrupt, os.Kill, syscall.SIGTERM)
 
-	stdlog.Println("데몬 시작")
-	errlog.Println("데몬 시작")
+	logger.Stdlog.Println("데몬 시작")
+	logger.Errlog.Println("데몬 시작")
 
 	// loop work cycle with accept connections or interrupt
 	// by system signal
 
 	var err error
 
-	connectionString, _ := getConnectionString(homedir)
-	errlog.Println("DB에 연결합니다 Connection String:", connectionString)
-
-	err = getAPIConfig(homedir, &apiconf)
+	err = readAPIConfig(basePath, &apiconf)
 	if err != nil {
-		panic(err)
+		logger.Stdlog.Fatal(err)
 	}
 
 	client = solapi.NewClient()
@@ -124,13 +103,10 @@ func (service *Service) Manage() (string, error) {
 		"Prefix":    apiconf.Prefix,
 	}
 
-	db, err = sql.Open("mysql", connectionString)
+	err = database.Connect(basePath)
 	if err != nil {
-		panic(err)
+		logger.Stdlog.Fatal(err)
 	}
-	db.SetConnMaxLifetime(time.Minute * 3)
-	db.SetMaxOpenConns(10)
-	db.SetMaxIdleConns(10)
 
 	go pollMsg()
 	go pollResult()
@@ -142,11 +118,11 @@ func (service *Service) Manage() (string, error) {
 	      runtime.GC()
 	      time.Sleep(time.Millisecond * 300)
 	      runtime.ReadMemStats(&rtm)
-	      errlog.Println("Allow:", rtm.Alloc)
-	      errlog.Println("TotalAllow:", rtm.TotalAlloc)
-	      errlog.Println("Sys:", rtm.Sys)
-	      errlog.Println("Mallocs:", rtm.Mallocs)
-	      errlog.Println("Frees:", rtm.Frees)
+	      logger.Errlog.Println("Allow:", rtm.Alloc)
+	      logger.Errlog.Println("TotalAllow:", rtm.TotalAlloc)
+	      logger.Errlog.Println("Sys:", rtm.Sys)
+	      logger.Errlog.Println("Mallocs:", rtm.Mallocs)
+	      logger.Errlog.Println("Frees:", rtm.Frees)
 	    }
 	  }()
 	*/
@@ -154,7 +130,7 @@ func (service *Service) Manage() (string, error) {
 	for {
 		select {
 		case killSignal := <-interrupt:
-			errlog.Println("시스템 시그널이 감지되었습니다:", killSignal)
+			logger.Errlog.Println("시스템 시그널이 감지되었습니다:", killSignal)
 			if killSignal == os.Interrupt {
 				return "Daemon was interrupted by system signal", nil
 			}
@@ -165,23 +141,11 @@ func (service *Service) Manage() (string, error) {
 	return usage, nil
 }
 
-func getConnectionString(homedir string) (string, error) {
+func readAPIConfig(homedir string, apiconf *APIConfig) error {
 	var b []byte
-	b, err := ioutil.ReadFile(homedir + "/db.json")
+	b, err := os.ReadFile(homedir + "/config.json")
 	if err != nil {
-		errlog.Println(err)
-		return "db.json 로딩 오류", err
-	}
-	_ = json.Unmarshal(b, &dbconf)
-	connectionString := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s", dbconf.User, dbconf.Password, dbconf.Host, dbconf.Port, dbconf.DBName)
-	return connectionString, nil
-}
-
-func getAPIConfig(homedir string, apiconf *APIConfig) error {
-	var b []byte
-	b, err := ioutil.ReadFile(homedir + "/config.json")
-	if err != nil {
-		errlog.Println(err)
+		logger.Errlog.Println(err)
 		return err
 	}
 	_ = json.Unmarshal(b, &apiconf)
@@ -191,9 +155,9 @@ func getAPIConfig(homedir string, apiconf *APIConfig) error {
 func pollMsg() {
 	for {
 		time.Sleep(time.Second * 1)
-		rows, err := db.Query("SELECT id, payload FROM msg WHERE sent = false AND scheduledAt <= NOW() AND scheduledAt >= SUBDATE(NOW(), INTERVAL 24 HOUR) AND sendAttempts < 3 LIMIT 10000")
+		rows, err := database.Db.Query("SELECT id, payload FROM msg WHERE sent = false AND scheduledAt <= NOW() AND scheduledAt >= SUBDATE(NOW(), INTERVAL 24 HOUR) AND sendAttempts < 3 LIMIT 10000")
 		if err != nil {
-			errlog.Println("[메시지발송] DB Query ERROR:", err)
+			logger.Errlog.Println("[메시지발송] DB Query ERROR:", err)
 			time.Sleep(time.Second * 5)
 			continue
 		}
@@ -214,7 +178,7 @@ func pollMsg() {
 				}
 				result, err := client.Messages.CreateGroup(params)
 				if err != nil {
-					errlog.Println(err)
+					logger.Errlog.Println(err)
 				}
 				groupId = result.GroupId
 			}
@@ -222,17 +186,17 @@ func pollMsg() {
 
 			err := rows.Scan(&id, &payload)
 			if err != nil {
-				errlog.Println(err)
+				logger.Errlog.Println(err)
 			}
 			_ = fmt.Sprintf("id: %u", id)
-			_, err = db.Exec("UPDATE msg SET sendAttempts = sendAttempts + 1 WHERE id = ?", id)
+			_, err = database.Db.Exec("UPDATE msg SET sendAttempts = sendAttempts + 1 WHERE id = ?", id)
 			if err != nil {
-				errlog.Println(err)
+				logger.Errlog.Println(err)
 				continue
 			}
 			err = json.Unmarshal([]byte(payload), &msgObj)
 			if err != nil {
-				errlog.Println(err)
+				logger.Errlog.Println(err)
 				continue
 			}
 
@@ -244,19 +208,19 @@ func pollMsg() {
 				if strings.HasPrefix(filename, "/") {
 					fullpath = filename
 				} else {
-					fullpath = homedir + "/files/" + filename
+					fullpath = basePath + "/files/" + filename
 				}
-				errlog.Println("파일 업로드:", fullpath)
+				logger.Errlog.Println("파일 업로드:", fullpath)
 				params := make(map[string]string)
 				params["file"] = fullpath
 				params["name"] = "customFileName"
 				params["type"] = "MMS"
 				result, err := client.Storage.UploadFile(params)
 				if err != nil {
-					errlog.Println(err)
+					logger.Errlog.Println(err)
 					continue
 				}
-				errlog.Println("이미지 아이디:", result.FileId)
+				logger.Errlog.Println("이미지 아이디:", result.FileId)
 				msgObj["imageId"] = result.FileId
 				delete(msgObj, "file")
 			}
@@ -271,7 +235,7 @@ func pollMsg() {
 
 			result, err := client.Messages.AddGroupMessage(groupId, msgParams)
 			if err != nil {
-				errlog.Println(err)
+				logger.Errlog.Println(err)
 				continue
 			}
 			for i, res := range result.ResultList {
@@ -279,16 +243,16 @@ func pollMsg() {
 				if res.StatusCode == "2000" {
 					status = "PENDING"
 				}
-				_, err = db.Exec("UPDATE msg SET result = json_object('messageId', ?, 'groupId', ?, 'status', ?, 'statusCode', ?, 'statusMessage', ?), sent = true WHERE id = ?", res.MessageId, groupId, status, res.StatusCode, res.StatusMessage, idList[i])
+				_, err = database.Db.Exec("UPDATE msg SET result = json_object('messageId', ?, 'groupId', ?, 'status', ?, 'statusCode', ?, 'statusMessage', ?), sent = true WHERE id = ?", res.MessageId, groupId, status, res.StatusCode, res.StatusMessage, idList[i])
 				if err != nil {
-					errlog.Println(err)
+					logger.Errlog.Println(err)
 					continue
 				}
 			}
 
 			groupInfo, err2 := client.Messages.SendGroup(groupId)
 			if err2 != nil {
-				errlog.Println(err2)
+				logger.Errlog.Println(err2)
 				continue
 			}
 			printObj(groupInfo.Count)
@@ -299,9 +263,9 @@ func pollMsg() {
 func pollLastReport() {
 	for {
 		time.Sleep(time.Second * 1)
-		rows, err := db.Query("SELECT id, messageId, statusCode FROM msg WHERE sent = true AND createdAt < SUBDATE(NOW(), INTERVAL 72 HOUR) AND status != 'COMPLETE' LIMIT 100")
+		rows, err := database.Db.Query("SELECT id, messageId, statusCode FROM msg WHERE sent = true AND createdAt < SUBDATE(NOW(), INTERVAL 72 HOUR) AND status != 'COMPLETE' LIMIT 100")
 		if err != nil {
-			errlog.Println("[마지막 리포트] DB Query ERROR:", err)
+			logger.Errlog.Println("[마지막 리포트] DB Query ERROR:", err)
 			time.Sleep(time.Second * 60)
 			continue
 		}
@@ -325,9 +289,9 @@ func pollLastReport() {
 func pollResult() {
 	for {
 		time.Sleep(time.Millisecond * 500)
-		rows, err := db.Query("SELECT id, messageId, statusCode FROM msg WHERE sent = true AND createdAt > SUBDATE(NOW(), INTERVAL 72 HOUR) AND updatedAt < SUBDATE(NOW(), INTERVAL (10 * (reportAttempts + 1)) SECOND) AND reportAttempts < 10 AND status != 'COMPLETE' LIMIT 100")
+		rows, err := database.Db.Query("SELECT id, messageId, statusCode FROM msg WHERE sent = true AND createdAt > SUBDATE(NOW(), INTERVAL 72 HOUR) AND updatedAt < SUBDATE(NOW(), INTERVAL (10 * (reportAttempts + 1)) SECOND) AND reportAttempts < 10 AND status != 'COMPLETE' LIMIT 100")
 		if err != nil {
-			errlog.Println("[리포트 처리] DB Query ERROR:", err)
+			logger.Errlog.Println("[리포트 처리] DB Query ERROR:", err)
 			time.Sleep(time.Second * 10)
 			continue
 		}
@@ -339,7 +303,7 @@ func pollResult() {
 		for rows.Next() {
 			_ = rows.Scan(&id, &messageId, &statusCode)
 
-			_, err = db.Exec("UPDATE msg SET reportAttempts = reportAttempts + 1, updatedAt = NOW() WHERE id = ?", id)
+			_, err = database.Db.Exec("UPDATE msg SET reportAttempts = reportAttempts + 1, updatedAt = NOW() WHERE id = ?", id)
 			messageIds = append(messageIds, messageId)
 		}
 		if len(messageIds) > 0 {
@@ -356,22 +320,22 @@ func syncMsgStatus(messageIds []string, statusCode string, defaultCode string) {
 	params["messageIds[in]"] = string(b)
 	params["limit"] = strconv.Itoa(len(messageIds))
 
-	errlog.Println("메시지 상태 동기화:", len(messageIds), "건")
+	logger.Errlog.Println("메시지 상태 동기화:", len(messageIds), "건")
 
 	result, err := client.Messages.GetMessageList(params)
 	if err != nil {
-		errlog.Println(err)
+		logger.Errlog.Println(err)
 	}
 
 	for _, res := range result.MessageList {
 		if res.StatusCode != statusCode {
-			_, err = db.Exec("UPDATE msg SET result = json_set(result, '$.status', ?, '$.statusCode', ?, '$.statusMessage', ?), updatedAt = NOW() WHERE messageId = ?", res.Status, res.StatusCode, res.Reason, res.MessageId)
+			_, err = database.Db.Exec("UPDATE msg SET result = json_set(result, '$.status', ?, '$.statusCode', ?, '$.statusMessage', ?), updatedAt = NOW() WHERE messageId = ?", res.Status, res.StatusCode, res.Reason, res.MessageId)
 			if err != nil {
 				panic(err)
 			}
 		} else {
 			if defaultCode != "" {
-				_, err = db.Exec("UPDATE msg SET result = json_set(result, '$.status', 'COMPLETE', '$.statusCode', ?, '$.statusMessage', ?), updatedAt = NOW() WHERE messageId = ?", defaultCode, "전송시간 초과", res.MessageId)
+				_, err = database.Db.Exec("UPDATE msg SET result = json_set(result, '$.status', 'COMPLETE', '$.statusCode', ?, '$.statusMessage', ?), updatedAt = NOW() WHERE messageId = ?", defaultCode, "전송시간 초과", res.MessageId)
 			}
 		}
 	}
@@ -384,56 +348,52 @@ func printObj(obj interface{}) {
 		panic(err)
 	}
 	msgStr := *(*string)(unsafe.Pointer(&msgBytes))
-	errlog.Println(msgStr)
+	logger.Errlog.Println(msgStr)
 }
 
 func init() {
 	agentHome := os.Getenv("AGENT_HOME")
 	if len(agentHome) > 0 {
-		homedir = agentHome
+		basePath = agentHome
 	}
 
-	// 콘솔 로그
-	stdlog = log.New(os.Stdout, "", log.Ldate|log.Ltime)
-
-	// 파일 이름 및 코드 라인 출력 시 다음 추가:  log.LstdFlags | log.Lshortfile
-	errlog = log.New(os.Stderr, "", log.Ldate|log.Ltime)
-	errlog.SetOutput(&lumberjack.Logger{
-		Filename:   homedir + "/logs/agent.log",
-		MaxSize:    500, // megabytes
-		MaxBackups: 3,
-		MaxAge:     28,   // days
-		Compress:   true, // disabled by default
-	})
+	logger.Create(basePath)
 }
 
 func main() {
+	dir, err := os.Getwd()
+	if err != nil {
+		logger.Stdlog.Fatal(err)
+	}
+	fmt.Printf("pwd: %s\n", dir)
+
 	daemonType := daemon.SystemDaemon
 	goos := runtime.GOOS
 	switch goos {
 	case "windows":
 		daemonType = daemon.SystemDaemon
-		stdlog.Println("Windows")
+		logger.Stdlog.Println("OS: Windows")
+		basePath = dir
 	case "darwin":
 		daemonType = daemon.UserAgent
-		stdlog.Println("MAC")
+		logger.Stdlog.Println("OS: MAC")
 	case "linux":
 		daemonType = daemon.SystemDaemon
-		stdlog.Println("Linux")
+		logger.Stdlog.Println("OS: Linux")
 	default:
 		fmt.Printf("%s.\n", goos)
 	}
 
 	srv, err := daemon.New(name, description, daemonType)
 	if err != nil {
-		errlog.Println("Error: ", err)
+		logger.Errlog.Println("Error: ", err)
 		os.Exit(1)
 	}
 	service := &Service{srv}
 	status, err := service.Manage()
 	if err != nil {
-		errlog.Println(status, "\nError: ", err)
+		logger.Errlog.Println(status, "\nError: ", err)
 		os.Exit(1)
 	}
-	stdlog.Println(status)
+	logger.Stdlog.Println(status)
 }
